tutorial_01/requirements.txt
django~=4.2.0

channels
daphne

django-pyhub-ai

django-environ

